Add `DistributionRef` type
