+++ src/server/os_trait.rs
@@ -3,13 +3,11 @@ use std::any::type_name;
 use std::path::PathBuf;
 use std::cmp::Ordering;
 
-use serde::{Deserialize, Serialize};
-use serde::{ser, de};
-
 use crate::server::install;
 use crate::server::detect::{VersionQuery, InstalledPackage, VersionResult};
 use crate::server::methods::{InstallationMethods, InstallMethod};
 use crate::server::version::Version;
+use crate::server::distribution::MajorVersion;
 use crate::server::init;
 
 
@@ -42,11 +40,6 @@ pub trait Method: fmt::Debug + Send + Sync {
         -> anyhow::Result<()>;
 }
 
-#[derive(PartialEq, PartialOrd, Ord, Eq, Debug, Clone)]
-pub enum MajorVersion {
-    Stable(Version<String>),
-    Nightly,
-}
 
 #[derive(Debug, Clone)]
 pub struct PreciseVersion {
@@ -92,44 +85,5 @@ impl PreciseVersion {
     }
     pub fn as_ver(&self) -> &Version<String> {
         &self.version
-    }
-}
-
-impl MajorVersion {
-    pub fn option(&self) -> String {
-        match self {
-            MajorVersion::Stable(v) => format!("--version={}", v.num()),
-            MajorVersion::Nightly => "--nightly".into(),
-        }
-    }
-    pub fn title(&self) -> &str {
-        match self {
-            MajorVersion::Stable(v) => v.num(),
-            MajorVersion::Nightly => "nightly",
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for MajorVersion {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where D: de::Deserializer<'de>,
-    {
-        let s: &str = Deserialize::deserialize(deserializer)?;
-        match s {
-            "nightly" => Ok(MajorVersion::Nightly),
-            s => Ok(MajorVersion::Stable(Version(s.into()))),
-        }
-    }
-}
-
-impl Serialize for MajorVersion {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: ser::Serializer,
-    {
-        serializer.serialize_str(match self {
-            MajorVersion::Stable(ver) => ver.num(),
-            MajorVersion::Nightly => "nightly",
-        })
     }
 }
